requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
ipykernel
bs4
pypdf
faiss-cpu
wikipedia
arxiv
langchainhub
flask
PyPDF2
python-multipart